+++ test/libsolidity/semanticTests/array/fixed_arrays_as_return_type.sol
@@ -21,6 +21,6 @@ contract B {
 // compileViaYul: also
 // ----
 // f() -> 2, 3, 4, 5, 6, 1000, 1001, 1002, 1003, 1004
-// gas irOptimized: 128110
+// gas irOptimized: 127910
 // gas legacy: 234719
 // gas legacyOptimized: 132639