+++ test/libsolidity/semanticTests/various/address_code.sol
@@ -17,7 +17,7 @@ contract C {
 // compileViaYul: also
 // ----
 // constructor() ->
-// gas irOptimized: 194717
+// gas irOptimized: 192317
 // gas legacy: 240889
 // gas legacyOptimized: 155314
 // initCode() -> 0x20, 0