+++ test/libsolidity/semanticTests/various/value_complex.sol
@@ -22,7 +22,7 @@ contract test {
 // compileViaYul: also
 // ----
 // constructor(), 20 wei ->
-// gas irOptimized: 190491
+// gas irOptimized: 190275
 // gas legacy: 265006
 // gas legacyOptimized: 182842
 // sendAmount(uint256): 5 -> 8